requirements.txt
streamlit
audio_recorder_streamlit
openai
python-dotenv
wave